data_clean_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-data-clean-env"
version = "1.0.0"
description = "Data Cleaning Environment for OpenEnv — real-world tabular data cleaning with RL"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "data_clean_env.server.app:main"

[tool.setuptools]
packages = ["data_clean_env", "data_clean_env.server"]

[tool.setuptools.package-dir]
data_clean_env = "."
"data_clean_env.server" = "server"

[tool.setuptools.package-data]
data_clean_env = ["**/*.yaml", "**/*.yml"]
